services/adaptive_narrative_engine/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
firebase-admin>=6.2.0
google-cloud-firestore>=2.13.0
google-cloud-pubsub>=2.18.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
openai>=1.0.0
google-generativeai>=0.3.0
google-cloud-secret-manager>=2.16.0
